+++ yummy-next/src/components/common/EnvScript.tsx
