+++ yummy-next/src/lib/client/auth/login/logOutHandler.ts
