+++ yummy-next/src/types/client/map/marker-clustering.ts
